FatihSahinWebsite/backend/requirements.txt
tensorflow
opencv-python-headless
numpy
fastapi
python-multipart
uvicorn
Pillow
